+++ matlab/dbqlsh_search.cpp
@@ -71,7 +71,7 @@ public:
             scanner.reset(quy + i * data.getDim());
             lsh.query(quy + i * data.getDim(), scanner);
             scanner.topk().genTopk();
-            std::vector<std::pair<float, unsigned> > tmp = scanner.topk().getTopk();
+            std::vector<std::pair<double, unsigned> > tmp = scanner.topk().getTopk();
             for (unsigned j = 0; j != tmp.size(); ++j)
             {
                 indices[i * K + j] = tmp[j].second;
